refactor: 프론트엔드 코드 정리
