+++ backend/src/main/java/com/multi/mariage/product/dto/request/ProductUpdateRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Data
 public class ProductUpdateRequest {
-    private String id;
+    private Long id;
     private String name;
     private String info;
     private double level;
