+++ frontend/src/apis/request/product.ts
@@ -2,23 +2,9 @@ import { API_PATH } from '@/constants/path';
 import { axiosWithAccessToken } from '../axios';
 import { ProductSaveType, ProductUpdateType } from '@/types/product';
 
-const requestSaveProduct = ({
-  name,
-  level,
-  info,
-  country,
-  upperCategory,
-  lowerCategory,
-  imageId,
-}: ProductSaveType) => {
+const requestSaveProduct = (productData: ProductSaveType) => {
   return axiosWithAccessToken.post(API_PATH.PRODUCT.SAVE, {
-    name,
-    level,
-    info,
-    country,
-    upperCategory,
-    lowerCategory,
-    imageId,
+    ...productData,
   });
 };
 
@@ -28,28 +14,9 @@ const requestProductInfo = async (productId: string) => {
   });
 };
 
-const requestUpdateProduct = async ({
-  id,
-  name,
-  info,
-  level,
-  country,
-  upperCategory,
-  lowerCategory,
-  imageId,
-  newImageId,
-}: ProductUpdateType) => {
-  console.log(newImageId);
+const requestUpdateProduct = async (productData: ProductUpdateType) => {
   return axiosWithAccessToken.patch(API_PATH.PRODUCT.UPDATE, {
-    id,
-    name,
-    info,
-    level,
-    country,
-    upperCategory,
-    lowerCategory,
-    imageId,
-    newImageId,
+    ...productData,
   });
 };
